references/CourseDesigner/requirements.txt
gradio
langgraph
typing
python-dotenv
#operator 
langgraph.checkpoint.sqlite
langchain-core
Langchain-openai